Make external URL configurable via environment variable
The externalURL function in the Handler has been updated to first check for the presence of an "EXTERNAL_URL" environment variable. If it is present, this value will be used, otherwise the function will fall back to the current behaviour of using the outbound IP and TCP port. Also, a minor correction in the nexusCache log message for cache hits, displaying correct info know.